[SPARK-58332][PYTHON][TEST] Move compare_or_generate_golden_matrix into GoldenFileTestMixin
### What changes were proposed in this pull request?
`compare_or_generate_golden_matrix` was duplicated verbatim across three PyArrow
golden-file test files. This moves it into `GoldenFileTestMixin`
(`python/pyspark/testing/goldenutils.py`), which all three suites already inherit,
and removes the local copies.

### Why are the changes needed?
Removes duplicated test machinery so the golden-file matrix driver has a single
implementation, following up on review feedback in #57435.

### Does this PR introduce _any_ user-facing change?
No. Test-only, behavior-preserving refactor.

### How was this patch tested?
Existing suites pass in compare mode (no golden files regenerated):
test_pyarrow_array_cast, test_pyarrow_arrow_to_pandas_default,
test_pyarrow_arrow_to_pandas_non_default.

### Was this patch authored or co-authored using generative AI tooling?
Generated-by: Claude Code (Opus 4.8)
